moved helper modules out of adapter package, references #23
